moved telemetry functions to specific file
