+++ price-feeder/oracle/provider/provider.go
@@ -5,8 +5,6 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/armon/go-metrics"
-	"github.com/cosmos/cosmos-sdk/telemetry"
 	"github.com/umee-network/umee/price-feeder/oracle/types"
 )
 
@@ -108,22 +106,4 @@ func checkHTTPStatus(resp *http.Response) error {
 		return fmt.Errorf("unexpected status: %s", resp.Status)
 	}
 	return nil
-}
-
-// telemetryWebsocketReconnect gives an standard way to add
-// `price_feeder_websocket_reconnect` metric.
-func telemetryWebsocketReconnect(n Name) {
-	telemetry.IncrCounterWithLabels(
-		[]string{
-			"websocket",
-			"reconnect",
-		},
-		1,
-		[]metrics.Label{
-			{
-				Name:  "provider",
-				Value: string(n),
-			},
-		},
-	)
 }
